refactorization of trace-analysis
Confirmed deletion of:
 	tools/trace-analysis/src/kieker/tools/trace/analysis/TraceAnalysisToolNewMain.java
