+++ tools/trace-analysis-gui/build.gradle
@@ -15,7 +15,7 @@ sourceSets.main.resources.srcDirs = [ '../resources' ]
 
 dependencies {
 	implementation project(":tools")
-	implementation project(":tools:trace-analysis")
+	implementation project(":tools:trace-analysis-old")
 
 	implementation "com.beust:jcommander:${jcommanderVersion}"
