+++ tools/trace-analysis-old/src/kieker/tools/trace/analysis/AnalysisTerminationException.java
